code/requirements.txt
boto3==1.9.188
botocore==1.12.188
google-cloud-monitoring==0.32.0
google-cloud-trace==0.22.0